Conflicts:
	.gitignore

+++ webrtc/modules/video_capture/include/video_capture_defines.h
@@ -104,7 +104,7 @@ class VideoCaptureExternal
 public:
     // |capture_time| must be specified in the NTP time format in milliseconds.
     virtual int32_t IncomingFrame(uint8_t* videoFrame,
-                                  int32_t videoFrameLength,
+                                  size_t videoFrameLength,
                                   const VideoCaptureCapability& frameInfo,
                                   int64_t captureTime = 0,
                                   bool faceDetected = false) = 0;
